### 2020-02-13 #2100 @whedon
[PRE REVIEW]: MaterialNet: A web-based graph explorer for materials science data
**Submitting author:** @waxlamp (<a href="http://orcid.org/0000-0003-3898-4626">Roni Choudhury</a>)
**Repository:** <a href="https://github.com/ToyotaResearchInstitute/materialnet" target ="_blank">https://github.com/ToyotaResearchInstitute/materialnet</a>
**Version:** 1.0.0
