### 2020-02-13 #2100 @danielskatz
👋 @waxlamp - we need an open source license in the repository.

Also, while I can see how this software is useful, can you explain how it is used in research? What problems does it help solve?
