### 2020-02-13 #2100 @danielskatz
You might want to add some of this to your paper and to your readme.
