### 2020-02-13 #2100 @danielskatz
👋 @dgasmith - would you be willing to review this for JOSS?
