### 2020-02-14 #2100 @waxlamp
>  @waxlamp - if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).

luthaf, rcannood, ati-ozgur, and sptennak all seem like good matches for our project. There's a mix of network science, materials science, and JavaScript expertise amongst them (though not concentrated in any one person).
