### 2020-02-14 #2100 @waxlamp
> You might want to add some of this to your paper and to your readme.

Thanks for this suggestion. I have folded some of that language into both the README and our paper text's introduction.
