### 2020-02-15 #2100 @danielskatz
👋 @luthaf, @rcannood, @ati-ozgur, @sptennak - would one or two of you be willing to review this submission to JOSS?
