### 2020-02-15 #2105 @danielskatz
👋 @Luthaf & @rcannood - thanks for agreeing to review - here's where the action happens.

Please read the comments above carefully, particularly in terms of accepting the invitation to be able to use the checkboxes, and in terms of not getting unwanted JOSS notifications.

Your job now is to examine the paper and the software, and to check off the items on your list that you can, and to explain what blocks you from checking off other items, either here if they are brief, or in new issues in the source repo if they are more complex. (If you do open issues in the source repo, please mention this review issue in those issues so that we can see them here as well.)

We expect that the review will be done in a couple of weeks, at most, or at least it will get to a list of blocking items in that timeframe.
