### 2020-02-24 #2105 @Luthaf
@danielskatz, are they JOSS guideline for documentation concerning SaaS/Online tools? Is user documentation/developer reference/separate installation information something that is required?
