### 2020-02-26 ToyotaResearchInstitute/materialnet#166 @rcannood
JOSS Review: Example usage
Please provide example usage in the documentation (README.md) on how to use MaterialNet (openjournals/joss-reviews#2105). For example, could you write a small markdown on how to perform a relevant analysis in this software? What information do you wish to obtain from MaterialNet?
