### 2020-02-26 ToyotaResearchInstitute/materialnet#166 @rcannood
Also, in the README.md it is not mentioned that the user should use `yarn start` to start the application.
