### 2020-02-26 ToyotaResearchInstitute/materialnet#167 @rcannood
JOSS Review: Local installation hangs
Related to the JOSS Review (openjournals/joss-reviews#2105). When I want to run the application locally, the application gets stuck loading a network.

![Screenshot from 2020-02-26 11-44-36](https://user-images.githubusercontent.com/553642/75337544-7b456580-588d-11ea-8bce-77e97d58e1b2.png)

These are the commands I entered in order to run the application:
```bash
git clone https://github.com/ToyotaResearchInstitute/materialnet.git
cd materialnet
yarn install
yarn start
```

Any ideas on how I can get this to work?
