### 2020-02-26 ToyotaResearchInstitute/materialnet#169 @rcannood
JOSS Review: Provide community guidelines
Please provide community guidelines for potential contributors (openjournals/joss-reviews#2105).

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
