### 2020-02-26 ToyotaResearchInstitute/materialnet#167 @waxlamp
Sorry about this. I will update the instructions and provide a data bundle suitable for local builds. I will report back here (via a pull request) when it's ready.
