### 2020-02-26 ToyotaResearchInstitute/materialnet#169 @waxlamp
@rcannood, please re-open this issue if you feel the CONTRIBUTING.md file is not sufficient.
