### 2020-02-26 ToyotaResearchInstitute/materialnet#167 @waxlamp
@rcannood, please try the build one more time using this branch: https://github.com/ToyotaResearchInstitute/materialnet/tree/data-bundle.

The README on that branch contains updated build instructions, and a pointer to a data bundle you can use to resolve this issue. If you run into trouble, please let me know and I will edit that branch until things are working.
