### 2020-02-26 ToyotaResearchInstitute/materialnet#166 @waxlamp
I am consulting with my co-authors and will add this information as soon as we have it.
