### 2020-02-26 #2105 @danielskatz
> are they JOSS guideline for documentation concerning SaaS/Online tools? Is user documentation/developer reference/separate installation information something that is required?

JOSS doesn't really provide detailed guidance here. You should probably consider this from the point of view of a user who wants to use the SaaS as well as a developer who wants to understand and potentially contribute to the underlying software.
