### 2020-02-27 #2105 @waxlamp
> > are they JOSS guideline for documentation concerning SaaS/Online tools? Is user documentation/developer reference/separate installation information something that is required?
>
> JOSS doesn't really provide detailed guidance here. You should probably consider this from the point of view of a user who wants to use the SaaS as well as a developer who wants to understand and potentially contribute to the underlying software.

Thanks for this insight. Creating detailed developer documentation is currently out of scope for our project, but of course I can see how it would enrich our open source offering here.
