### 2020-02-27 #2105 @danielskatz
I think that's a reasonable answer, as long as the community guidelines match - expectations for community contributions will depend on developer documentation to some extent.
