### 2020-03-02 ToyotaResearchInstitute/materialnet#166 @Luthaf
The example usage should also contains a description on how one could go from a database of materials to a full visualization: how to create the network, how to save it (which format(s) are accepted by MaterialNet), and how to load it in MaterialNet.
