### 2020-03-03 ToyotaResearchInstitute/materialnet#166 @waxlamp
Please see [this branch](https://github.com/ToyotaResearchInstitute/materialnet/tree/example-usage) for an updated README that addresses these issues, and let me know what you think!
