### 2020-03-04 #2105 @rcannood
Would be nice if @whedon could also generate a diff :)

[Here](https://github.com/ToyotaResearchInstitute/materialnet/compare/470318edf5b9234c1dab672061aa1c11da8b8145...master#diff-16d3fe279379b2a19e836e62ab840c12) are the changes w.r.t. the original proof.
