### 2020-03-04 ToyotaResearchInstitute/materialnet#166 @rcannood
Alright, good job!
