### 2020-03-06 #2105 @danielskatz
👋 @waxlamp - please check the references, particular the cases (upper vs lower) for title and journal names. Add {}s in the .bib file to protect cases through compilation, and then use ```@whedon generate pdf``` to generate a new PDF.  The immediate thing I see is that voronoi should be Voronoi, but please check for anything else as well.
