### 2020-03-06 #2105 @danielskatz
I've also created https://github.com/ToyotaResearchInstitute/materialnet/pull/175 with a few more changes in the paper, and a question about wording.
