### 2020-03-09 #2105 @waxlamp
Here's the requested information:
- tagged release version: v1.0.0 (https://github.com/ToyotaResearchInstitute/materialnet/releases/tag/v1.0.0)
- Zenodo DOI: http://doi.org/10.5281/zenodo.3701759

Please let me know if these look ok.
