### 2020-03-09 #2105 @waxlamp
> Thanks to @Luthaf and @rcannood for reviewing, and congratulations to @waxlamp and coauthors!

Thank you, reviewers! And thank you @danielskatz for your support during this very first JOSS paper submission experience for me!
